repare import of PopGenome data

+++ R/sum_stat_generate.R
@@ -26,7 +26,8 @@ generateSumStats <- function(files, program, parameters, dm, seg_sites) {
   }
 
   # Add the parameters of the simulation
-  sum_stats <- list(pars=parameters)
+  sum_stats <- list()
+  if (!missing(parameters)) sum_stats[['pars']] <- parameters
 
   # Add seg_sites
   if ('seg.sites' %in% model_stats) {
@@ -49,11 +50,13 @@ generateSumStats <- function(files, program, parameters, dm, seg_sites) {
   }
   
   # Add files if needed
-  if ('file' %in% model_stats) {
-    sum_stats[['file']] <- files
-  } else {
-    unlink(unlist(files))
+  if (!missing(files)) {
+    if ('file' %in% model_stats) {
+      sum_stats[['file']] <- files
+    } else {
+      unlink(unlist(files))
+    }
   }
-  
+
   sum_stats
 }
